+++ src/components/FeaturedContent/FeaturedContent.stories.tsx
@@ -4,11 +4,11 @@ import Button from "../Button/Button";
 import Heading from "../Heading/Heading";
 import Text from "../Text/Text";
 import SimpleGrid from "../Grid/SimpleGrid";
-
 import FeaturedContent, {
   featuredContentWidthArray,
   featuredContentPositionArray,
 } from "./FeaturedContent";
+import { getPlaceholderImage } from "../../utils/utils";
 
 const meta: Meta<typeof FeaturedContent> = {
   title: "Components/Basic Elements/FeaturedContent",
@@ -56,7 +56,7 @@ export const WithControls: Story = {
     imageProps: undefined,
     "imageProps.alt": "Alt text",
     "imageProps.position": "end",
-    "imageProps.src": "//loremflickr.com/600/600/new+york+public+library",
+    "imageProps.src": getPlaceholderImage(),
     "imageProps.width": "default",
     isFullWidth: undefined,
     textContent: undefined,
@@ -125,7 +125,7 @@ export const LayoutVariations: Story = {
           alt: "Alt text",
           width: "default",
           position: "end",
-          src: "//loremflickr.com/600/600/new+york+public+library",
+          src: getPlaceholderImage(),
         }}
       />
       <FeaturedContent
@@ -148,7 +148,7 @@ export const LayoutVariations: Story = {
           alt: "Alt text",
           width: "default",
           position: "end",
-          src: "//loremflickr.com/600/600/new+york+public+library",
+          src: getPlaceholderImage(),
         }}
       />
     </SimpleGrid>
@@ -176,7 +176,7 @@ export const ImagePositionVariations: Story = {
           alt: "Alt text",
           width: "default",
           position: "start",
-          src: "//loremflickr.com/600/600/new+york+public+library",
+          src: getPlaceholderImage(),
         }}
       />
       <FeaturedContent
@@ -197,7 +197,7 @@ export const ImagePositionVariations: Story = {
           alt: "Alt text",
           width: "default",
           position: "end",
-          src: "//loremflickr.com/600/600/new+york+public+library",
+          src: getPlaceholderImage(),
         }}
       />
     </SimpleGrid>
@@ -225,7 +225,7 @@ export const imageWidthVariations: Story = {
           alt: "Alt text",
           width: "oneQuarter",
           position: "end",
-          src: "//loremflickr.com/600/600/new+york+public+library",
+          src: getPlaceholderImage(),
         }}
       />
       <FeaturedContent
@@ -246,7 +246,7 @@ export const imageWidthVariations: Story = {
           alt: "Alt text",
           width: "oneThird",
           position: "end",
-          src: "//loremflickr.com/600/600/new+york+public+library",
+          src: getPlaceholderImage(),
         }}
       />
       <FeaturedContent
@@ -267,7 +267,7 @@ export const imageWidthVariations: Story = {
           alt: "Alt text",
           width: "oneHalf",
           position: "end",
-          src: "//loremflickr.com/600/600/new+york+public+library",
+          src: getPlaceholderImage(),
         }}
       />
       <FeaturedContent
@@ -288,7 +288,7 @@ export const imageWidthVariations: Story = {
           alt: "Alt text",
           width: "twoThirds",
           position: "end",
-          src: "//loremflickr.com/600/600/new+york+public+library",
+          src: getPlaceholderImage(),
         }}
       />
       <FeaturedContent
@@ -309,7 +309,7 @@ export const imageWidthVariations: Story = {
           alt: "Alt text",
           width: "threeQuarters",
           position: "end",
-          src: "//loremflickr.com/600/600/new+york+public+library",
+          src: getPlaceholderImage(),
         }}
       />
     </SimpleGrid>
@@ -336,7 +336,7 @@ export const textContentVariations: Story = {
           alt: "Alt text",
           width: "threeQuarters",
           position: "end",
-          src: "//loremflickr.com/600/600/new+york+public+library",
+          src: getPlaceholderImage(),
         }}
       />
       <FeaturedContent
@@ -355,7 +355,7 @@ export const textContentVariations: Story = {
           alt: "Alt text",
           width: "oneHalf",
           position: "end",
-          src: "//loremflickr.com/600/600/new+york+public+library",
+          src: getPlaceholderImage(),
         }}
       />
     </SimpleGrid>