+++ src/components/Tooltip/Tooltip.test.tsx
@@ -11,6 +11,7 @@ import { Button } from "../Button/Button";
 import { Icon } from "../Icons/Icon";
 import { Image } from "../Image/Image";
 import { Tooltip, TooltipProps } from "./Tooltip";
+import { getPlaceholderImage } from "../../utils/utils";
 
 const buttonLabel = "Hover me";
 const tooltipContent =
@@ -40,12 +41,7 @@ describe("Tooltip accessibility", () => {
   it("passes axe accessibility test with Image content", async () => {
     render(
       <Tooltip
-        content={
-          <Image
-            src="//loremflickr.com/300/300/new+york+public+library"
-            alt=""
-          />
-        }
+        content={<Image src={getPlaceholderImage("smaller", 0)} alt="" />}
       >
         {buttonLabel}
       </Tooltip>